Blank line cull
Unfortunately, each {% ... %} Jekyll code causes a blank line in the
output HTML. To remove lots of blank lines from the source code, I've
put lots of the Jekyll code on one line. I also loop over the correct
number of pages. (5 instead of 10 - this will need to be changed if a
page is added).