Navigation order now handles in a csv data file
This code is a bit nicer than previous. Removes the need for a navorder
tag in the front matter of each page. Also removes the need for
pages_list, which required slightly strange code. To change navigation
order we just need to update navigation.csv with the correct page titles
- which should be defined in the front matter of each individual page.