Bug fixes
Wait for the ajax load to finish before saving the content to the state.
Fixes issues where if the ajax request took too long the user would see
the content pop in. It would also store the wrong content to the state,
causing issues when the back button is pressed.

Also, when the back button was pressed, the link in the top nav that was
pressed to change page is selected, and focus would make this grey. I.e.
if you went from index to research and then clicked the back button, the
research button that you clicked (on index) would be grey when you went
back to index. This has been fixed by having :focus being the standard
colours.