Ensure all content is loaded.
In 5c3cc610ae025837b27841871410a77a1807a99b we fixed an issue where some
of the content on index was not loaded when the link was clicked. This
was because of the #content > * in the ajax request. We fix this by
added a superfluous div, and getting this div in the ajax request.